# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onfly - Pedido de Viagens",
        "description": "Serviço que gerencia os pedidos de viagens dos clientes",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        },
        {
            "url": "http://my-default-host.com"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Gera um token para autenticação de usuário",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "description": "Dados do usuário para criação e autenticação",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha_do_usuario"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "senha_do_usuario"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário registrado e token gerado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João da Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "usuario@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token_gerado_aqui"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo email é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação de dados falhou.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo email é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro inesperado ao criar o usuário e gerar o token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro inesperado ao criar o usuário."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login do usuário e gera um token de autenticação",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "description": "Credenciais do usuário para autenticação",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha_do_usuario"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "token_gerado_aqui"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo email é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro inesperado ao gerar o token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro inesperado ao gerar o token."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilize um token Bearer para autenticação nas rotas protegidas da API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}